+++ geo_knowledge_hub/theme/assets/semantic-ui/js/geo_knowledge_hub_detail/index.js
@@ -26,6 +26,7 @@ const componentsAvailable = [
   "record_management",
   "related_links",
   "related_packages",
+  "request_training",
 ];
 
 renderFactory(componentsAvailable);